Completed Game interface and model to fetch games by p1_id and p2_id. Have not yet finished gameHistory route

+++ twodoku-server/src/main/java/io/twodoku/twodokuserver/repository/UserInterface.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.twodoku.twodokuserver.models.User;
 
-public interface UserInterface extends JpaRepository<User, Long> {
+public interface UserInterface extends JpaRepository<User, Integer> {
   User findByName(String name);
   User findByEmail(String email);
 }
